refactor[#186] : FSD 구조에 맞게 SettingItems 분할
