+++ src/entities/profile/ui/index.ts
@@ -1,4 +1,3 @@
 export * from "./TemporaryMarkingBar";
 export * from "./GalleryGrid";
 export * from "./ProfileOverview";
-export * from "./SettingItems";
