+++ src/entities/setting/ui/VersionInfo.tsx
@@ -1,0 +1,10 @@
+const SettingClassName =
+  "py-2 flex justify-between w-full text-grey-700 title-2";
+
+// 버전 정보
+export const VersionInfo = () => (
+  <div className={SettingClassName}>
+    <p>버전 정보</p>
+    <span className="text-grey-700 body-2">v0.1</span>
+  </div>
+);
