+++ src/features/setting/constants/index.ts
@@ -1,0 +1,2 @@
+export const SettingClassName =
+  "py-2 flex justify-between w-full text-grey-700 title-2 hover:bg-grey-50 focus-visible:bg-grey-50 focus-visible:outline-none active:bg-grey-100";
