+++ src/features/setting/ui/index.ts
@@ -1,0 +1,7 @@
+export * from "./EditMyInfo";
+export * from "./AccountManagement";
+export * from "./MyActivity";
+export * from "./Notification";
+export * from "./TermOfService";
+export * from "./Report";
+export * from "./LogoutButton";
